# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation Ch5",
    "description": "API Documentation for all tables",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:2000",
      "description": "Local server"
    }
  ],
  "tags": [
    {
      "name": "Register",
      "description": "Regis User"
    },
    {
      "name": "Login",
      "description": "Get token"
    },
    {
      "name": "Users",
      "description": "Manage table users & profiles"
    },
    {
      "name": "BankAccounts",
      "description": "Manage table bank_accounts"
    },
    {
      "name": "Transactions",
      "description": "Manage table transactions"
    }
  ],
  "paths": {
    "/api/v1/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Login user",
        "description": "Mengembalikan token JWT jika login berhasil.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Berhasil login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid email or password"
          }
        }
      }
    },
    "/api/v1/users/register": {
      "post": {
        "tags": [
          "Register"
        ],
        "summary": "Create a new user and profile",
        "description": "Create a new user and associated profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userData": {
                    "$ref": "#/components/schemas/User"
                  },
                  "profileData": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User and profile created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users",
        "description": "Retrieve a list of all users",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch"
          }
        }
      }
    },
    "/api/v1/users/:{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get detail users",
        "description": "Retrieve a list of all users",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID user"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDetail"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Perbarui user dan profil berdasarkan ID",
        "description": "Memperbarui data user dan profil berdasarkan userId. Diperlukan autentikasi Bearer Token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID user"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userData": {
                    "$ref": "#/components/schemas/User"
                  },
                  "profileData": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User dan profil berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Gagal memperbarui data user dan profil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Hapus user berdasarkan ID",
        "description": "Menghapus user berdasarkan userId. Diperlukan autentikasi Bearer Token.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID user"
          }
        ],
        "responses": {
          "200": {
            "description": "User berhasil dihapus",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Gagal menghapus user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts": {
      "post": {
        "tags": [
          "BankAccounts"
        ],
        "summary": "Create a new bank account",
        "description": "Create a new bank account for a",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer",
                    "example": 1
                  },
                  "bankName": {
                    "type": "string",
                    "example": "Bank XYZ"
                  },
                  "bankAccountNumber": {
                    "type": "string",
                    "example": 1234567890
                  },
                  "balance": {
                    "type": "number",
                    "example": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bank account created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "BankAccounts"
        ],
        "summary": "Get all bank accounts",
        "description": "Retrieve a list of all bank accounts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of bank accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BankAccount"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch accounts"
          }
        }
      }
    },
    "/api/v1/accounts/{accountId}": {
      "get": {
        "tags": [
          "BankAccounts"
        ],
        "summary": "Get detail users",
        "description": "Retrieve a list of all users",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID account"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountDetail"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch"
          }
        }
      }
    },
    "/api/v1/transactions": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Perform a transaction",
        "description": "Transfer money from one account to",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sourceAccountId": {
                    "type": "integer",
                    "example": 1
                  },
                  "destinationAccountId": {
                    "type": "integer",
                    "example": 2
                  },
                  "amount": {
                    "type": "number",
                    "example": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction successful"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get all transactions",
        "description": "Retrieve a list of all transactions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch transactions"
          }
        }
      }
    },
    "/api/v1/transactions/{transactionId}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get detail transactions",
        "description": "Retrieve a list of all transactions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID transaction"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "amount": {
                      "type": "integer"
                    },
                    "sourceAccount": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "userId": {
                          "type": "integer"
                        },
                        "bankName": {
                          "type": "string"
                        },
                        "bankAccountNumber": {
                          "type": "string"
                        },
                        "balance": {
                          "type": "number"
                        }
                      }
                    },
                    "destinationAccount": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "userId": {
                          "type": "integer"
                        },
                        "bankName": {
                          "type": "string"
                        },
                        "bankAccountNumber": {
                          "type": "string"
                        },
                        "balance": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch transactions"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "Profile": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "userId": {
                "type": "integer"
              },
              "identityType": {
                "type": "string"
              },
              "identityNumber": {
                "type": "string"
              },
              "addres": {
                "type": "string"
              }
            }
          }
        }
      },
      "AccountDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "bank name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "bankAccountNumber": {
            "type": "integer"
          },
          "balance": {
            "type": "integer"
          },
          "User": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Profile": {
        "type": "object",
        "properties": {
          "identityType": {
            "type": "string"
          },
          "identityNumber": {
            "type": "string"
          },
          "addres": {
            "type": "string"
          }
        }
      },
      "BankAccount": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "bankName": {
            "type": "string"
          },
          "bankAccountNumber": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "sourceAccountId": {
            "type": "integer"
          },
          "destinationAccountId": {
            "type": "integer"
          },
          "amount": {
            "type": "number"
          }
        }
      }
    }
  }
}